Move quadrature logic to its own header file
